melanoma_cancer/melanoma_cancer/requirements.txt
absl-py==2.1.0
appdirs==1.4.4
astor==0.8.1
attrs==23.2.0
boto3==1.34.23
botocore==1.34.23
Brotli==1.1.0
brotlipy==0.7.0
build==1.0.3
cachetools==5.3.3
clyent==1.2.1
constantly==15.1.0
cssselect==1.1.0
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.5
datashape==0.5.2
decorator==4.4.2
distlib==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
faiss-cpu==1.8.0
fake-useragent==1.4.0
feather-format==0.4.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.25.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.2
httpx==0.26.0
huepy==1.2.1
inflate64==1.0.0
inflection==0.5.1
instabot==0.117.0
jsonpatch==1.33
jsonpointer==2.1
keras==2.15.0
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.48
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langsmith==0.1.52
libclang==18.1.1
marshmallow==3.21.1
ml-dtypes==0.2.0
mock==5.1.0
MouseInfo==0.1.3
moviepy==1.0.3
mplcursors==0.5.3
multivolumefile==0.2.3
munkres==1.1.4
namex==0.0.8
nh3==0.2.15
oauthlib==3.2.2
openai==1.24.1
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
outcome==1.3.0.post0
packaging==23.2
patsy==0.5.3
pep8==1.7.1
pip-tools==7.3.0
ply==3.11
proglog==0.1.10
protobuf==4.25.3
py7zr==0.21.0
pyarrow==11.0.0
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pybcj==1.0.2
pycryptodomex==3.20.0
PyDispatcher==2.0.5
PyGetWindow==0.0.9
pyls-spyder==0.4.0
PyMsgBox==1.0.9
#pyobjc-core==10.2
# pyobjc-framework-Cocoa==10.2
# pyobjc-framework-Quartz==10.2
pypdf==4.2.0
pyperclip==1.8.2
pyppmd==1.1.0
pyproject_hooks==1.0.0
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
pytweening==1.2.0
pyzstd==0.15.10
queuelib==1.5.0
readme-renderer==42.0
reliability==0.8.16
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.7.0
rsa==4.9
rubicon-objc==0.4.7
s3fs==0.4.2
s3transfer==0.10.0
schedule==1.2.1
selenium==4.19.0
sklearn-pandas==2.2.0
sniffio==1.3.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
texttable==1.7.0
tiktoken==0.6.0
torchtuples==0.2.2
trio==0.25.0
trio-websocket==0.11.1
twine==4.0.2
typing-inspect==0.9.0
typing_extensions==4.10.0
virtualenv==20.25.0
webdriver-manager==4.0.1
webencodings==0.5.1
#wsproto==1.2.0
